Do not convert an Ion Frames to an Exit frames.
This patch just flag its previous frame as a bailed JS frame, such as we
know that we have to walk the exit frame as-if it had the size of a
JS frame.  This is a temporary hack before bug 717297 can be done.

This is needed by fun.apply implementation since fun.apply use the frame
descriptor to recover the number of arguments it push on the stack.